# prueba/openapi.yaml
openapi: 3.0.3
info:
  title: Devsu Prueba API
  description: API para gestión de clientes, cuentas y movimientos bancarios
  version: 1.0.0
  contact:
    name: Devsu
servers:
  - url: http://localhost:8080
    description: Servidor local

tags:
  - name: Clients
    description: Operaciones relacionadas con clientes
  - name: Accounts
    description: Operaciones relacionadas con cuentas bancarias
  - name: Movements
    description: Operaciones relacionadas con movimientos bancarios

paths:
  /api/v1/clients:
    get:
      tags:
        - Clients
      summary: Obtener todos los clientes
      description: Retorna una lista de todos los clientes registrados
      operationId: getClients
      responses:
        '200':
          description: Lista de clientes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetClientDto'
    post:
      tags:
        - Clients
      summary: Crear un nuevo cliente
      description: Crea un nuevo cliente en el sistema
      operationId: postClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostClientDto'
      responses:
        '201':
          description: Cliente creado exitosamente
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/clients/{id}:
    get:
      tags:
        - Clients
      summary: Obtener un cliente por ID
      description: Retorna un cliente específico por su identificador único
      operationId: getClientById
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetClientDto'
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Clients
      summary: Actualizar un cliente
      description: Actualiza la información de un cliente existente
      operationId: putClient
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutClientDto'
      responses:
        '204':
          description: Cliente actualizado exitosamente
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Clients
      summary: Eliminar un cliente
      description: Elimina un cliente del sistema
      operationId: deleteClient
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cliente eliminado exitosamente
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/accounts:
    get:
      tags:
        - Accounts
      summary: Obtener todas las cuentas
      description: Retorna una lista de todas las cuentas bancarias registradas
      operationId: getAccounts
      responses:
        '200':
          description: Lista de cuentas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetAccountDto'
    post:
      tags:
        - Accounts
      summary: Crear una nueva cuenta
      description: Crea una nueva cuenta bancaria en el sistema
      operationId: postAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostAccountDto'
      responses:
        '201':
          description: Cuenta creada exitosamente
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Obtener una cuenta por ID
      description: Retorna una cuenta específica por su identificador único
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único de la cuenta
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAccountDto'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Accounts
      summary: Actualizar una cuenta
      description: Actualiza la información de una cuenta existente
      operationId: putAccount
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único de la cuenta
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutAccountDto'
      responses:
        '204':
          description: Cuenta actualizada exitosamente
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/movements/{clientId}:
    get:
      tags:
        - Movements
      summary: Obtener movimientos por rango de fechas y cliente
      description: Retorna una lista de movimientos filtrados por rango de fechas y cliente
      operationId: getMovements
      parameters:
        - name: clientId
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: string
            format: uuid
        - name: init_date
          in: query
          required: true
          description: Fecha de inicio del rango de búsqueda
          schema:
            type: string
            format: date-time
          example: "2026-01-01T00:00:00"
        - name: end_date
          in: query
          required: true
          description: Fecha de fin del rango de búsqueda
          schema:
            type: string
            format: date-time
          example: "2026-01-31T23:59:59"
      responses:
        '200':
          description: Lista de movimientos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetMovementDto'
        '400':
          description: Parámetros de consulta inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/movements:
    post:
      tags:
        - Movements
      summary: Crear un nuevo movimiento
      description: Registra un nuevo movimiento bancario (débito o crédito)
      operationId: postMovement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMovementDto'
      responses:
        '201':
          description: Movimiento creado exitosamente
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    GetClientDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del cliente
        name:
          type: string
          description: Nombre completo del cliente
          example: "Juan Pérez"
        identification:
          type: string
          description: Número de identificación del cliente
          example: "1234567890"
        address:
          type: string
          description: Dirección del cliente
          example: "Calle Principal 123"
        phone:
          type: string
          description: Número de teléfono del cliente
          example: "0987654321"
        gender:
          $ref: '#/components/schemas/Gender'
        password:
          type: string
          description: Contraseña del cliente
        status:
          $ref: '#/components/schemas/Status'

    PostClientDto:
      type: object
      required:
        - identification
        - address
        - phone
        - gender
        - password
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del cliente (opcional)
        name:
          type: string
          description: Nombre completo del cliente
          example: "Juan Pérez"
        identification:
          type: string
          description: Número de identificación del cliente
          pattern: '^[0-9]+$'
          minLength: 10
          maxLength: 20
          example: "1234567890"
        address:
          type: string
          description: Dirección del cliente
          example: "Calle Principal 123"
        phone:
          type: string
          description: Número de teléfono del cliente
          pattern: '^[0-9]+$'
          example: "0987654321"
        gender:
          $ref: '#/components/schemas/Gender'
        password:
          type: string
          description: Contraseña del cliente
          example: "password123"
        status:
          $ref: '#/components/schemas/Status'

    PutClientDto:
      type: object
      properties:
        name:
          type: string
          description: Nombre completo del cliente
          example: "Juan Pérez"
        identification:
          type: string
          description: Número de identificación del cliente
          example: "1234567890"
        address:
          type: string
          description: Dirección del cliente
          example: "Calle Principal 123"
        phone:
          type: string
          description: Número de teléfono del cliente
          example: "0987654321"
        gender:
          $ref: '#/components/schemas/Gender'
        password:
          type: string
          description: Contraseña del cliente
        status:
          $ref: '#/components/schemas/Status'

    GetAccountDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único de la cuenta
        number:
          type: string
          description: Número de la cuenta bancaria
          example: "123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          type: number
          format: float
          description: Saldo inicial de la cuenta
          example: 1000.00
        balance:
          type: number
          format: float
          description: Saldo actual de la cuenta
          example: 1500.50
        status:
          $ref: '#/components/schemas/Status'

    PostAccountDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único de la cuenta (opcional)
        clientId:
          type: string
          format: uuid
          description: Identificador del cliente propietario de la cuenta
        number:
          type: string
          description: Número de la cuenta bancaria
          example: "123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          type: number
          format: float
          description: Saldo inicial de la cuenta
          example: 1000.00
        balance:
          type: number
          format: float
          description: Saldo actual de la cuenta
          example: 1000.00
        status:
          $ref: '#/components/schemas/Status'

    PutAccountDto:
      type: object
      properties:
        number:
          type: string
          description: Número de la cuenta bancaria
          example: "123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          type: number
          format: float
          description: Saldo inicial de la cuenta
          example: 1000.00
        balance:
          type: number
          format: float
          description: Saldo actual de la cuenta
          example: 1500.50
        status:
          $ref: '#/components/schemas/Status'

    GetMovementDto:
      type: object
      properties:
        date:
          type: string
          format: date-time
          description: Fecha y hora del movimiento
          example: "2026-01-21T10:30:00"
        client:
          type: string
          description: Nombre del cliente
          example: "Juan Pérez"
        numberAccount:
          type: string
          description: Número de la cuenta asociada
          example: "123456789"
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          type: number
          format: float
          description: Saldo inicial antes del movimiento
          example: 1000.00
        status:
          $ref: '#/components/schemas/Status'
        movement:
          type: number
          format: float
          description: Valor del movimiento (positivo para créditos, negativo para débitos)
          example: 500.00
        availableBalance:
          type: number
          format: float
          description: Saldo disponible después del movimiento
          example: 1500.00

    PostMovementDto:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
          description: Identificador de la cuenta en la que se realiza el movimiento
        date:
          type: string
          format: date-time
          description: Fecha y hora del movimiento
          example: "2026-01-21T10:30:00"
        movementType:
          $ref: '#/components/schemas/MovementType'
        value:
          type: number
          format: float
          description: Valor del movimiento
          example: 500.00

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de error
          example: "Recurso no encontrado"
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora del error
        status:
          type: integer
          description: Código de estado HTTP
          example: 404

    AccountType:
      type: string
      enum:
        - SAVINGS
        - CURRENT
      description: Tipo de cuenta bancaria

    Gender:
      type: string
      enum:
        - MALE
        - FEMALE
      description: Género del cliente

    MovementType:
      type: string
      enum:
        - DEBIT
        - CREDIT
      description: Tipo de movimiento bancario

    Status:
      type: string
      enum:
        - ACTIVATE
        - INACTIVATE
      description: Estado de activación