# prueba/src/main/java/com/devsu/prueba/service/dto/PostMovementDto.java
package com.devsu.prueba.service.dto;

import com.devsu.prueba.entities.enums.MovementType;
import lombok.*;

import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.UUID;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class PostMovementDto {
    private UUID accountId;
    private LocalDateTime date;
    private MovementType movementType;
    private BigDecimal value;
}