# prueba/src/main/java/com/devsu/prueba/service/dto/PutAccountDto.java
package com.devsu.prueba.service.dto;

import java.math.BigDecimal;

import com.devsu.prueba.entities.enums.AccountType;
import com.devsu.prueba.entities.enums.Status;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class PutAccountDto {
    private String number;
    private AccountType accountType;
    private BigDecimal initialBalance;
    private BigDecimal balance;
    private Status status;
}